# Conflicts:
#	src/main/kotlin/io/github/ezrnest/symbolic/ExprCal.kt
#	src/main/kotlin/io/github/ezrnest/symbolic/ExprContext.kt
#	src/main/kotlin/io/github/ezrnest/symbolic/Main.kt
#	src/main/kotlin/io/github/ezrnest/symbolic/Node.kt
#	src/main/kotlin/io/github/ezrnest/symbolic/SimRule.kt
#	src/main/kotlin/io/github/ezrnest/symbolic/SimUtils.kt

+++ src/main/kotlin/io/github/ezrnest/model/Models.kt
@@ -459,7 +459,7 @@ class DoubleAsReals(
         return kotlin.math.sqrt(x)
     }
 
-    override fun nroot(x: Double, n: Long): Double {
+    override fun nroot(x: Double, n: Int): Double {
         return x.pow(1.0 / n)
     }
 
@@ -588,7 +588,7 @@ class BigDecimalAsReals(val mc: MathContext = MathContext.DECIMAL128) : Reals<Bi
         }
     }
 
-    override fun nroot(x: BigDecimal, n: Long): BigDecimal {
+    override fun nroot(x: BigDecimal, n: Int): BigDecimal {
         TODO("Not yet implemented")
     }
